% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{bailen}
\alias{bailen}
\title{El texto del EN "Bailén"}
\format{Vector de caracteres con 6101 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=Bail\%C3\%A9n&oldid=196724}
}
\usage{
bailen
}
\description{
Un dataset que contiene el texto "Bailén", cuarto Episodio Nacional
de Benito Pérez Galdós, publicado en 1873. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
