% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{batalla_arapiles}
\alias{batalla_arapiles}
\title{El texto del EN "La batalla de los Arapiles"}
\format{Vector de caracteres con 10254 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=La_batalla_de_los_Arapiles&oldid=228663}
}
\usage{
batalla_arapiles
}
\description{
Un dataset que contiene el texto "Juan Martín el Empecinado", décimo Episodio Nacional
de Benito Pérez Galdós, publicado en 1875. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
