% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{cadiz}
\alias{cadiz}
\title{El texto del EN "Cádiz"}
\format{Vector de caracteres con 8846 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=C\%C3\%A1diz&oldid=196671}
}
\usage{
cadiz
}
\description{
Un dataset que contiene el texto "Cádiz", octavo Episodio Nacional
de Benito Pérez Galdós, publicado en 1874. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
