% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{corte_carlos_iv}
\alias{corte_carlos_iv}
\title{El texto del EN "La corte de Carlos IV"}
\format{Vector de caracteres con 7326 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=La_corte_de_Carlos_IV&oldid=502348}
}
\usage{
corte_carlos_iv
}
\description{
Un dataset que contiene el texto "La corte de Carlos IV", esegundo Episodio Nacional
de Benito Pérez Galdós, publicado en 1873. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
