% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{empecinado}
\alias{empecinado}
\title{El texto del EN "Juan Martín el Empecinado"}
\format{Vector de caracteres con 7040 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=Juan_Mart\%C3\%ADn_el_Empecinado&oldid=196677}
}
\usage{
empecinado
}
\description{
Un dataset que contiene el texto "Juan Martín el Empecinado", noveno Episodio Nacional
de Benito Pérez Galdós, publicado en 1874. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
