% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{package}
\name{galdosenr}
\alias{galdosenr}
\alias{galdosenr-package}
\title{Episodios nacionales de Benito Pérez Galdós}
\description{
Este paquete contiene los textos completos de los Episodios Nacionales
de la serie 1.ª publicados entre 1873 y 1875 y formateados para
un correcto análisis textual. Los seis primeros elementos de cada
vector son el título, el autor y el número del episodios separados
por un elemento en vacío.
}
