% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{marzo19_2mayo}
\alias{marzo19_2mayo}
\title{El texto del EN "El 19 de marzo y el 2 de mayo"}
\format{Vector de caracteres con 6737 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=El_19_de_marzo_y_el_2_de_mayo&oldid=502350}
}
\usage{
marzo19_2mayo
}
\description{
Un dataset que contiene el texto "El 19 de marzo y el 2 de mayo", tercer Episodio Nacional
de Benito Pérez Galdós, publicado en 1873. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
