% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{napoleon_en_chamartin}
\alias{napoleon_en_chamartin}
\title{El texto del EN "Napoleón en Chamartín"}
\format{Vector de caracteres con 7868 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=Napole\%C3\%B3n_en_Chamart\%C3\%ADn&oldid=196683}
}
\usage{
napoleon_en_chamartin
}
\description{
Un dataset que contiene el texto "Napoleón en Chamartín", quinto Episodio Nacional
de Benito Pérez Galdós, publicado en 1874. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
