% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{trafalgar}
\alias{trafalgar}
\title{El texto del EN "Trafalgar"}
\format{Vector de caracteres con 4853 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=Trafalgar&oldid=196685}
}
\usage{
trafalgar
}
\description{
Un dataset que contiene el texto "Trafalgar", el primer Episodio Nacional
de Beneito Pérez Galdós, publicado en 1873. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
