% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdosenr-package.r
\docType{data}
\name{zaragoza}
\alias{zaragoza}
\title{El texto del EN "Zaragoza"}
\format{Vector de caracteres con 6472 elementos.}
\source{
\url{https://es.wikisource.org/w/index.php?title=Zaragoza&oldid=196687}
}
\usage{
zaragoza
}
\description{
Un dataset que contiene el texto "Zaragoza", sexto Episodio Nacional
de Benito Pérez Galdós, publicado en 1874. El texto en codificación UTF-8 se
obtuvo de Wikisource, se revisó y corrigió. Está dividido en elementos
de 80 caracteres. Los párrafos están separados por un elemento en blanco.
}
\keyword{datasets}
